% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{decomposition_step}
\alias{decomposition_step}
\title{Decomposition Step}
\usage{
decomposition_step(exp_files)
}
\arguments{
\item{exp_files}{\itemize{
\item list of expression files for each context
}}
}
\value{
returns the shared component of expression per individual and each specific expression component for each of the C contexts
}
\description{
Function to decompose expression into one shared component and specific components per context to run crocotel direct
}
