% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_equicor_chol}
\alias{make_equicor_chol}
\title{Build the Cholesky factor of an equicorrelation matrix}
\usage{
make_equicor_chol(n_contexts, rho)
}
\arguments{
\item{n_contexts}{Integer. Dimension of the matrix.}

\item{rho}{Numeric. Off-diagonal correlation.}
}
\value{
Lower-triangular matrix (n_contexts x n_contexts).
}
\description{
Sigma_cc = 1, Sigma_cc' = rho (c != c').
Returns the lower-triangular Cholesky factor L such that Sigma = L L^T.
Valid for rho in (-1/(n_contexts - 1), 1).
}
\keyword{internal}
