% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardise_cols}
\alias{standardise_cols}
\title{Standardise columns of a matrix to zero mean and unit variance}
\usage{
standardise_cols(X)
}
\arguments{
\item{X}{Numeric matrix.}
}
\value{
Matrix with standardised columns. Columns with zero variance
  are left as-is with a warning.
}
\description{
Standardise columns of a matrix to zero mean and unit variance
}
\keyword{internal}
