% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GReXs.R
\name{create_GReXs}
\alias{create_GReXs}
\title{Creates GReXs (Genetically regulated predictors of expression) for one gene across contexts}
\usage{
create_GReXs(
  X_file,
  exp_files,
  contexts,
  out_dir,
  gene_name,
  context_thresh = 3,
  alpha = 0.5,
  num_folds = 10,
  run_GBAT = FALSE
)
}
\arguments{
\item{X_file}{\itemize{
\item a genotype file with the set of individuals in all Y_files and their corresponding cis-SNPs. Individuals are row names, no column names
}}

\item{exp_files}{\itemize{
\item vector with list of all expression files per context. The individuals (row names) must be a subset of genotype (X_file). Contains an unnamed
}}

\item{contexts}{\itemize{
\item vector of strings which have context names in them.
}}

\item{out_dir}{\itemize{
\item output directory for GReXs
}}

\item{gene_name}{\itemize{
\item identifier for current gene being run. Add this prefix to all the saved results files... necessary to distinguish results where more than one gene-analysis is run.
}}

\item{context_thresh}{\itemize{
\item minimum number of contexts to run C-STEM on.
}}

\item{alpha}{\itemize{
\item The regularization constant. Default is .5 (eNet). Minimum value of 1e-4.
}}

\item{num_folds}{\itemize{
\item Number of folds for cross-validation
}}

\item{run_GBAT}{\itemize{
\item Takes values of TRUE or FALSE. Default is FALSE, if TRUE then will run the GBAT* method too.
}}
}
\value{
writes out a file of predicted expression across individuals and contexts
}
\description{
Creates GReXs (Genetically regulated predictors of expression) for one gene across contexts
}
