% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crocotel_SNP_based.R
\name{crocotel_SNP_based}
\alias{crocotel_SNP_based}
\title{crocotel SNP based}
\usage{
crocotel_SNP_based(
  genotype_file,
  snpsloc_file,
  exp_files,
  geneloc_file,
  out_dir,
  output_file_name_cis = file.path(out_dir, paste0("crocotel_SNP_based.cis_pairs.txt")),
  output_file_name_tra = file.path(out_dir, paste0("crocotel_SNP_based.trans_pairs.txt")),
  method = "MatrixEQTL",
  use_model = modelLINEAR,
  cis_dist = 1e+06,
  pv_threshold_cis = 0,
  pv_threshold_tra = 0.05,
  error_covariance = numeric()
)
}
\arguments{
\item{out_dir}{Output directory.}

\item{output_file_name_cis}{Path to write cis-eQTL output.}

\item{output_file_name_tra}{Path to write trans-eQTL output.}

\item{method}{Either "MatrixEQTL" or "tensorqtl".}

\item{use_model}{MatrixEQTL model (default: modelLINEAR).}

\item{cis_dist}{Distance threshold for cis-eQTLs.}

\item{pv_threshold_cis}{P-value threshold for cis.}

\item{pv_threshold_tra}{P-value threshold for trans.}

\item{error_covariance}{Covariance matrix (or numeric()).}

\item{SNP_file_name}{Full path to SNP genotype matrix.}

\item{snps_location_file_name}{Full path to SNP location file.}

\item{expression_file_name}{Full path to expression matrix.}

\item{gene_location_file_name}{Full path to gene location file.}

\item{context}{Context name for labeling output.}
}
\value{
Writes cis-eQTLs and trans-eQTLs (optional) to file.
}
\description{
Function to map SNP-based trans eQTLs - cis window is defined as 1Mb
}
