% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Creates GReXs (Genetically regulated predictors of expression) for one gene across contexts}
\usage{
format_data(
  exp_files,
  geneloc_file,
  snpsloc_file,
  genotypes_file,
  out_dir,
  cis_window = 1e+06
)
}
\arguments{
\item{exp_files}{\itemize{
\item vector of file names for each context's expression for all genes and individuals
}}
}
\value{
writes out a file of predicted expression across individuals and contexts
}
\description{
Creates GReXs (Genetically regulated predictors of expression) for one gene across contexts
}
